+++ src/main/java/michaelscott/command/UnmarkCommand.java
@@ -1,6 +1,6 @@
 package michaelscott.command;
 
-import michaelscott.MichaelScottException;
+import michaelscott.utils.MichaelScottException;
 import michaelscott.task.Task;
 import michaelscott.task.TaskList;
