+++ src/main/java/michaelscott/task/TaskList.java
@@ -1,7 +1,7 @@
 package michaelscott.task;
 
 import java.util.ArrayList;
-import michaelscott.MichaelScottException;
+import michaelscott.utils.MichaelScottException;
 
 /**
  * This class represents the list used to store all the tasks
