+++ src/main/java/michaelscott/utils/MichaelScottException.java
@@ -1,4 +1,4 @@
-package michaelscott;
+package michaelscott.utils;
 
 public class MichaelScottException extends Exception {
     public MichaelScottException(String errMessage) {
